+++ core/run/ScriptUtils.cpp
@@ -198,8 +198,10 @@ void Chai::registerBindings(chaiscript::ChaiScript& chai) {
     chai.add(chaiscript::fun(&Sph::cos), "cos");
     chai.add(chaiscript::fun(&Sph::sin), "sin");
     chai.add(chaiscript::fun(&Sph::lerp<double, double>), "lerp");
-    chai.add(chaiscript::fun(&Sph::abs<double>), "abs");
+    chai.add(chaiscript::fun([](double x) { return x > 0 ? x : -x; }), "abs");
     chai.add(chaiscript::fun(&Sph::pow<double>), "pow");
+    chai.add(chaiscript::fun(&Sph::exp), "exp");
+    chai.add(chaiscript::fun(&Sph::log), "log");
 
     // vector utils
     chai.add(chaiscript::user_type<Vec3>(), "Vec3");