+++ core/system/Settings.cpp
@@ -1,6 +1,7 @@
 #include "gravity/AggregateSolver.h"
 #include "io/Output.h"
 #include "system/Settings.impl.h"
+#include "physics/Eos.h"
 
 NAMESPACE_SPH_BEGIN
 
@@ -604,7 +605,7 @@ const RunSettings& getDefaultSettings() {
         "Specifies how the density is evolved. Can be one of the following:\n" + EnumMap::getDesc<ContinuityEnum>() },
     { RunSettingsId::SPH_DISCRETIZATION,            "sph.discretization",              DiscretizationEnum::STANDARD,
         "Specifies a discretization of SPH equations. Can be one of the following:\n" + EnumMap::getDesc<DiscretizationEnum>() },
-    { RunSettingsId::SPH_STABILIZATION_DAMPING,     "sph.stabilization_damping",       0.1_f,
+    { RunSettingsId::SPH_STABILIZATION_DAMPING,     "sph.stabilization_damping",       0.01_f,
         "Specifies the damping coefficient of particle velocities. This is mainly intended for stabilization phase, "
        "it should not be used in the main simulation." },
 
@@ -770,6 +771,7 @@ static RegisterEnum<EosEnum> sEos({
         "body, can be used to simulate "
         "dust interacting only by friction or gravity." },
     { EosEnum::IDEAL_GAS, "ideal_gas", "Equation of state for ideal gas." },
+    { EosEnum::POLYTROPIC, "polytropic", "Polytropic equation of state." },
     { EosEnum::TAIT, "tait", "Tait equation of state for simulations of liquids." },
     { EosEnum::MIE_GRUNEISEN,
         "mie_gruneisen",
@@ -777,6 +779,7 @@ static RegisterEnum<EosEnum> sEos({
     { EosEnum::TILLOTSON, "tillotson", "Tillotson equation of stats." },
     { EosEnum::MURNAGHAN, "murnaghan", "Murnaghan equation of state." },
     { EosEnum::SIMPLIFIED_TILLOTSON, "simplified_tillotson", "Simplified version of the Tillotson equation."},
+    { EosEnum::HUBBARD_MACFARLANE, "hubbard_macfarlane", "Equation of state from Hubbard & MacFarlane (1980)."},
     { EosEnum::ANEOS,
         "aneos",
         "ANEOS equation of state, requires look-up table of values for given material." },
@@ -819,6 +822,10 @@ const BodySettings& getDefaultSettings() {
         "Slope of the Hugoniot curve, used in Mie-Gruneisen EoS." },
     { BodySettingsId::BULK_SOUND_SPEED,        "eos.mie_gruneises.bulk_sound_speed", 3933._f,  // value for copper taken from wikipedia
         "Bulk sound speed used in Mie-Gruneisen EoS." },
+    { BodySettingsId::POLYTROPIC_CONSTANT,     "eos.polytrope.constant",       2e6_f,
+        "Constant K used by the polytropic equation of state" },
+    { BodySettingsId::HUBBARD_MACFARLANE_TYPE, "eos.hubbard_macfarlane.type", HubbardMacFarlaneEos::Type::ROCK,
+        "Type of material used by the Hubbard & MacFarlane EoS." },
 
     /// Yielding & Damage
     { BodySettingsId::RHEOLOGY_YIELDING,    "rheology.yielding",            YieldingEnum::VON_MISES,