+++ core/system/Settings.h
@@ -1398,7 +1398,13 @@ enum class EosEnum {
     SIMPLIFIED_TILLOTSON,
 
     /// ANEOS given by look-up table
-    ANEOS
+    ANEOS,
+
+    /// Polytropic EoS
+    POLYTROPIC,
+
+    /// Equation of state from Hubbard & MacFarlane (1980).
+    HUBBARD_MACFARLANE,
 };
 
 /// \brief Settings of a single body / gas phase / ...
@@ -1520,6 +1526,12 @@ enum class BodySettingsId {
 
     /// Linear Hugoniot slope coefficient used in Mie-Gruneisen equation of state
     HUGONIOT_SLOPE = 32,
+
+    /// Constant for the polytropic equation of state.
+    POLYTROPIC_CONSTANT = 76,
+
+    /// Material type used by the Hubbard & MacFarlane EoS
+    HUBBARD_MACFARLANE_TYPE = 77,
 
     /// Model of stress reducing used within the rheological model
     RHEOLOGY_YIELDING = 33,