+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2022-10-25
+- added a node for stabilization by adjusting the particle densities
+
+## 2021-07-30
+- added render overlay to surface and volume renderers
+
+## 2021-06-25
+- added attractors to the surface renderer
+
 ## 2021-05-08
 - added a solver based on position based dynamics